admin-settle-add-validation
Now if someone who is not a solver or admin tries to settle an order, they will get CantDo. If they are a solver or admin and the order is not in dispute, mostro will notified that they cannot do because of the status, and if it is not in dispute but they did not taken it, will notified correctly.